+++ packages/apptive_grid_form/lib/src/widgets/form_widget/attachment_form_widget.dart
@@ -81,14 +81,14 @@ class _AttachmentFormWidgetState extends State<AttachmentFormWidget>
         initialValue: widget.component.data,
         builder: (formState) {
           return InputDecorator(
-            decoration: widget.component.baseDecoration.copyWith(
-              errorText: formState.errorText,
-              contentPadding: EdgeInsets.zero,
-              border: InputBorder.none,
-              errorBorder: InputBorder.none,
-              isDense: true,
-              filled: false,
-            ),
+            decoration: widget.component.baseDecoration(context).copyWith(
+                  errorText: formState.errorText,
+                  contentPadding: EdgeInsets.zero,
+                  border: InputBorder.none,
+                  errorBorder: InputBorder.none,
+                  isDense: true,
+                  filled: false,
+                ),
             child: Column(
               mainAxisSize: MainAxisSize.min,
               crossAxisAlignment: CrossAxisAlignment.stretch,