+++ packages/apptive_grid_form/lib/src/widgets/form_widget/enum_form_widget.dart
@@ -34,12 +34,12 @@ class _EnumFormWidgetState extends State<EnumFormWidget>
         enabled: widget.component.enabled,
         builder: (fieldState) {
           return InputDecorator(
-            decoration: widget.component.baseDecoration.copyWith(
-              errorText: fieldState.errorText,
-              contentPadding: EdgeInsets.zero,
-              filled: false,
-              border: InputBorder.none,
-            ),
+            decoration: widget.component.baseDecoration(context).copyWith(
+                  errorText: fieldState.errorText,
+                  contentPadding: EdgeInsets.zero,
+                  filled: false,
+                  border: InputBorder.none,
+                ),
             child: Column(
               crossAxisAlignment: CrossAxisAlignment.stretch,
               mainAxisSize: MainAxisSize.min,
@@ -95,7 +95,7 @@ class _EnumFormWidgetState extends State<EnumFormWidget>
       validator: _validate,
       autovalidateMode: AutovalidateMode.onUserInteraction,
       initialValue: widget.component.data.value,
-      decoration: widget.component.baseDecoration,
+      decoration: widget.component.baseDecoration(context),
     );
   }